# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Paste API server. It is used to publish, read, edit text posts.",
        "title": "Paste API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/pastes/": {
            "get": {
                "description": "Retrieves a paste from the database by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pastes"
                ],
                "summary": "Retrieve a paste",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of the paste",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Category ID of the paste",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order, e.g., -created_at",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved paste",
                        "schema": {
                            "$ref": "#/definitions/service.ListPastesOutput"
                        }
                    },
                    "422": {
                        "description": "Unprocessing data",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests, rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new paste in the database by input data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pastes"
                ],
                "summary": "Create a new paste",
                "parameters": [
                    {
                        "description": "Paste creation input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreatePasteInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created paste",
                        "schema": {
                            "$ref": "#/definitions/service.PasteResp"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "URL of the newly created paste"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable data",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests, rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/pastes/{id}": {
            "get": {
                "description": "Retrieves a paste from the database by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pastes"
                ],
                "summary": "Retrieve a paste",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Paste ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved paste",
                        "schema": {
                            "$ref": "#/definitions/service.PasteResp"
                        }
                    },
                    "404": {
                        "description": "Paste not found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests, rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a paste from the database by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pastes"
                ],
                "summary": "Deletes a paste",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Paste ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted paste"
                    },
                    "403": {
                        "description": "User is not allowed to edit this paste",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Paste not found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests, rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the paste in the database by ID and input data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pastes"
                ],
                "summary": "Update the paste",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Paste ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Paste update input",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/service.UpdatePasteInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated paste",
                        "schema": {
                            "$ref": "#/definitions/service.PasteResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User is not allowed to edit this paste",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable data",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests, rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/pastes/{id}/permission/{user_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gives a permission to user in the url to update / delete paste in the url.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pastes",
                    "users"
                ],
                "summary": "Gives a write permission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Paste ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully gave permission",
                        "schema": {
                            "$ref": "#/definitions/service.PastePermissionResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests, rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Metadata": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "first_page": {
                    "type": "integer"
                },
                "last_page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
        "models.Paste": {
            "type": "object",
            "properties": {
                "can_edit": {
                    "type": "boolean"
                },
                "category": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "models.Permission": {
            "type": "object",
            "properties": {
                "paste_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "service.CreatePasteInput": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "integer"
                },
                "minutes": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "service.ListPastesOutput": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/models.Metadata"
                },
                "pastes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Paste"
                    }
                }
            }
        },
        "service.PastePermissionResponse": {
            "type": "object",
            "properties": {
                "permission": {
                    "$ref": "#/definitions/models.Permission"
                }
            }
        },
        "service.PasteResp": {
            "type": "object",
            "properties": {
                "paste": {
                    "$ref": "#/definitions/models.Paste"
                }
            }
        },
        "service.UpdatePasteInput": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "integer"
                },
                "minutes": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "To authorize, use the format: Bearer \u003cyour_token\u003e. For example: Bearer abc123xyz",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}